requirements.txt
Flask==3.0.0
Jinja2>=3.1.2
Werkzeug>=3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
elevenlabs>=2.0.0
requests==2.31.0

